feat(homenavigation): adjust dynamic import, skeleton ui to homenavigation
