+++ service/app/src/widgets/navigation/ui/homeNavigation/index.ts
@@ -1,0 +1,1 @@
+export { HomeNavigation } from "./homeNavigation"
